move search to separate component and tighten regex
